fix wrong names for fixtures
