+++ cypress/fixtures/family.ts
@@ -1,6 +1,5 @@
 import { faker } from '@faker-js/faker';
 
-export const projectFixture = {
-  name: () => faker.lorem.word(100),
-  description: () => faker.lorem.lines(1),
+export const familyFixture = {
+  name: () => faker.name.fullName(),
 };
