+++ cypress/fixtures/project.ts
@@ -1,5 +1,6 @@
 import { faker } from '@faker-js/faker';
 
-export const familyFixture = {
-  name: () => faker.name.fullName(),
+export const projectFixture = {
+  name: () => faker.lorem.word(100),
+  description: () => faker.lorem.lines(1),
 };
